requirements.txt
Django==2.2.5
django-cors-headers==3.1.1
djangorestframework==3.10.3
djangorestframework-stubs==1.0.0
pytz==2019.2
sqlparse==0.3.0
